api_yamdb/requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.0.12
Django==3.2
django-filter==22.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
idna==3.4
importlib-metadata==4.3.0
iniconfig==2.0.0
gunicorn==20.0.4
packaging==23.0
pluggy==0.13.1
psycopg2-binary
py==1.11.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2022.7.1
requests==2.26.0
sqlparse==0.4.3
toml==0.10.2
typing-extensions==4.4.0
urllib3==1.26.14
zipp==3.11.0